added filters in the map
