+++ applications/vessel-history/public/locales/source/translations.json
@@ -82,6 +82,7 @@
     "noHighlitingSettingsMessage": "You've not defined the highlight settings yet. Please click Settings to configure the activity of interest."
   },
   "map": {
+    "filters": "Filter events",
     "loading": "Map loading",
     "toggleLayers": "Toggle layers"
   },
