+++ applications/vessel-history/src/features/event-filters/EventFilters.module.css
@@ -1,0 +1,8 @@
+.filterSelector {
+  display: flex;
+  align-items: center;
+}
+
+.filterSwitch {
+  margin-right: 10px;
+}
