+++ applications/vessel-history/src/features/map/map.slice.ts
@@ -10,7 +10,7 @@ import {
 import { RootState } from 'store'
 import { selectUrlMapZoomQuery } from 'routes/routes.selectors'
 import { Range } from 'types'
-import { selectFilters } from 'features/profile/filters/filters.slice'
+import { selectFilters } from 'features/event-filters/filters.slice'
 
 export interface MapState {
   generatorsConfig: AnyGeneratorConfig[]
