+++ applications/vessel-history/src/features/vessels/activity/vessels-activity.selectors.test.ts
@@ -1,11 +1,11 @@
-import { Filters } from 'features/profile/filters/filters.slice'
+import { Filters } from 'features/event-filters/filters.slice'
 import { selectFilteredEvents, RenderedEvent } from './vessels-activity.selectors'
 import {
   loiteringAndEncounterEvents,
   portVisitEvents,
 } from './__mocks__/selectEventsWithRenderingInfo.mock'
 
-jest.mock('features/profile/filters/filters.slice')
+jest.mock('features/event-filters/filters.slice')
 
 describe('selectFilteredEvents', () => {
   const cases: [string, string, string, RenderedEvent[], Filters, number][] = [
